+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Enhanced issue display with GitHub project information
+
 ## [2.0.0] - 2025-07-25
 
 ### Added